feat: creation of model order_book, updating endpoints user, books, cart e purchase

+++ src/services/User/DeleteUserService.ts
@@ -4,10 +4,11 @@ import AppError from "../../errors/AppError";
 
 interface Request {
   id: string;
+  idLogged: string;
 }
 
 export default class DeleteUserService {
-  public async execute({ id }: Request): Promise<DeleteResult> {
+  public async execute({ id, idLogged }: Request): Promise<DeleteResult> {
     const userRepository = getCustomRepository(UserRepository);
 
     const user = await userRepository.findOne({
@@ -15,9 +16,25 @@ export default class DeleteUserService {
         id,
       },
     });
+
     if (!user) {
-      throw new AppError("Not found any user with this id");
+      throw new AppError("Not found any user with this id.");
     }
+    
+    const userLogged = await userRepository.findOne({
+      where: {
+        id: idLogged,
+      },
+    });
+
+    if (!user) {
+      throw new AppError("Not found any user with this id.");
+    }   
+
+    if (id !== idLogged && userLogged?.is_adm === false) {
+      throw new AppError("Missing admin permissions", 401)
+    }
+
     return userRepository.delete(id);
   }
 }
